+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-06-28 22:46:45
-// Commit: feat(i18n): add async i18n stream provider
+// Updated at: 2026-06-30 19:10:52
+// Commit: feat(i18n): implement shimmer loading skeleton in i18n
 
 class i18nComponentState {
-  final String commitRef = '2401';
-  final String lastMessage = 'feat(i18n): add async i18n stream provider';
-  final DateTime timestamp = DateTime.parse('2026-06-28T22:46:45+05:00');
+  final String commitRef = '2434';
+  final String lastMessage = 'feat(i18n): implement shimmer loading skeleton in i18n';
+  final DateTime timestamp = DateTime.parse('2026-06-30T19:10:52+05:00');
 }